Address review feedback for VM lifecycle operations
Use built-in VM API methods (vm.wait_for_shutdown, vm.reboot)
instead of custom implementations. Remove unnecessary wrapper
functions test_reboot and test_reset from lifecycle_base.
Move check_network_access to provider/sriov/check_points to
deduplicate code between iommu_device_lifecycle and
iommu_device_settings. Replace eval() with ast.literal_eval()
for safer parameter parsing.

Reference: LIBVIRTAT-22036
Made-with: Cursor
